Modified bug fix in AzurePowerShell.ps1
I modified the fix per advice from reviewer niadak.

+++ Tasks/AzurePowerShell/AzurePowerShell.ps1
@@ -83,7 +83,9 @@ try {
         }
 }
 finally {
-    if ($__vstsAzPSInlineScriptPath -and (Test-Path -LiteralPath $__vstsAzPSInlineScriptPath) ) {
-        Remove-Item -LiteralPath $__vstsAzPSInlineScriptPath -ErrorAction 'SilentlyContinue'
+    If ($scriptType -eq "InlineScript") {
+        if ($__vstsAzPSInlineScriptPath -and (Test-Path -LiteralPath $__vstsAzPSInlineScriptPath) ) {
+            Remove-Item -LiteralPath $__vstsAzPSInlineScriptPath -ErrorAction 'SilentlyContinue'
+        }
     }
 }
